[refactor] Res 파일 각 feature 모듈 별로 관리하는 방식으로 refactoring
다국어 지원을 하는 경우가 아니라면, 개발의 편의성 증진을 위해 각 모듈에서 사용하는 res 파일은 해당 모듈이 가지고 있는게 합리적

DroidKnights, nia 도 같은 방식을 따르고 있음
