+++ feature/mate/src/main/kotlin/com/tripmate/android/mate/viewmodel/MateUiState.kt
@@ -17,7 +17,7 @@ data class MateUiState(
                         "Searching Title 1",
                         "강원도 춘천시",
                         "This is the description for item 1",
-                        R.drawable.img_camera_with_flash,
+                        R.drawable.ic_mypage,
                         lat = 37.517235,
                         lon = 127.047325,
                         isSearching = true,
@@ -26,7 +26,7 @@ data class MateUiState(
                         "Searching Title 2",
                         "강원도 춘천시",
                         "This is the description for item 2",
-                        R.drawable.img_camera_with_flash,
+                        R.drawable.ic_mypage,
                         lat = 37.5,
                         lon = 127.2,
                     ),
@@ -34,7 +34,7 @@ data class MateUiState(
                         "Searching Title 3",
                         "강원도 춘천시",
                         "This is the description for item 3",
-                        R.drawable.img_camera_with_flash,
+                        R.drawable.ic_mypage,
                         lat = 37.5,
                         lon = 127.3,
                         isSearching = true,
@@ -48,7 +48,7 @@ data class MateUiState(
                         "Title 1",
                         "강원도 춘천시",
                         "This is the description for item 1",
-                        R.drawable.img_camera_with_flash,
+                        R.drawable.ic_mypage,
                         lat = 37.5,
                         lon = 127.0,
                         isSearching = true,
@@ -57,7 +57,7 @@ data class MateUiState(
                         "Title 2",
                         "강원도 춘천시",
                         "This is the description for item 2",
-                        R.drawable.img_camera_with_flash,
+                        R.drawable.ic_mypage,
                         lat = 37.6,
                         lon = 127.0,
                     ),
@@ -65,7 +65,7 @@ data class MateUiState(
                         "Title 3",
                         "강원도 춘천시",
                         "This is the description for item 3",
-                        R.drawable.img_camera_with_flash,
+                        R.drawable.ic_mypage,
                         lat = 37.7,
                         lon = 127.0,
                         isSearching = true,
