+++ feature/personalization/src/main/kotlin/com/tripmate/android/feature/personalization/component/UnderAgeDialog.kt
@@ -2,9 +2,9 @@ package com.tripmate.android.feature.personalization.component
 
 import androidx.compose.runtime.Composable
 import com.tripmate.android.core.designsystem.ComponentPreview
-import com.tripmate.android.core.designsystem.R
 import com.tripmate.android.core.designsystem.component.TripmateDialog
 import com.tripmate.android.core.designsystem.theme.TripmateTheme
+import com.tripmate.android.feature.personalization.R
 
 @Composable
 fun UnderAgeDialog(
