+++ feature/personalization/src/main/res/drawable/img_camera_with_flash.png
