+++ feature/personalization/src/main/res/drawable/img_ferris_wheel.png
